refactor graph queries folder, update paths
