+++ apps/charge-network-service/src/common/constants/graph-queries/fuse-blocks.ts
@@ -1,0 +1,16 @@
+import { gql } from 'graphql-request'
+
+export const getBlock = gql`
+    query blockQuery($timestampFrom: Int, $timestampTo: Int) {
+        blocks(
+            first: 1
+            where: { timestamp_gt: $timestampFrom, timestamp_lt: $timestampTo }
+            orderBy: timestamp
+            orderDirection: desc
+        ) {
+            id
+            number
+            timestamp
+        }
+    }
+`
