+++ apps/charge-network-service/src/common/constants/index.ts
@@ -1,0 +1,6 @@
+export const voltageProtocolFee = 0.0005
+
+export const daysInYear = 365
+
+// Providers
+export const voltBarId = 'volt-bar'
