--- apps/charge-network-service/src/common/constants/staking-providers.ts
@@ -1,1 +1,0 @@
-export const voltBarId = 'volt-bar'
